+++ src/fascicleReaders.h
@@ -1,18 +1,18 @@
 #ifndef _FIBERREADERS_H
 #define _FIBERREADERS_H
 
-#include <Rcpp.h>
+#include <cpp11.hpp>
 #include <string>
 #include <vtkPolyData.h>
 #include <vtkSmartPointer.h>
 
-void WriteCSV(const vtkSmartPointer<vtkPolyData> &inputData,
-              std::string &outputFile);
-// [[Rcpp::export]]
-void ReadVTK(const std::string &inputTracts, std::string &outputFile);
-// [[Rcpp::export]]
-void ReadVTP(const std::string &inputTracts, std::string &outputFile);
-// [[Rcpp::export]]
-void ReadFDS(const std::string &inputTracts, std::string &outputFile);
+cpp11::writable::list
+PolyDataToList(const vtkSmartPointer<vtkPolyData> &polyData);
+[[cpp11::register]]
+cpp11::writable::list ReadVTK(const std::string &inputTracts);
+[[cpp11::register]]
+cpp11::writable::list ReadVTP(const std::string &inputTracts);
+[[cpp11::register]]
+cpp11::writable::list ReadFDS(const std::string &inputTracts);
 
 #endif /* _FIBERREADERS_H */