+++ src/fascicleWriters.h
@@ -1,18 +1,18 @@
 #ifndef _FIBERWRITERS_H
 #define _FIBERWRITERS_H
 
-#include <Rcpp.h>
+#include <cpp11.hpp>
 #include <string>
 #include <vtkPolyData.h>
 #include <vtkSmartPointer.h>
 
-void ReadCSV(const std::string &inputFile,
-             vtkSmartPointer<vtkPolyData> &inputData);
-// [[Rcpp::export]]
-void WriteVTK(const std::string &inputTracts, std::string &outputFile);
-// [[Rcpp::export]]
-void WriteVTP(const std::string &inputTracts, std::string &outputFile);
-// [[Rcpp::export]]
-void WriteFDS(const std::string &inputTracts, std::string &outputFile);
+void ListToPolyData(cpp11::list inputData,
+                    vtkSmartPointer<vtkPolyData> &polyData);
+[[cpp11::register]]
+void WriteVTK(cpp11::list inputTracts, const std::string &outputFile);
+[[cpp11::register]]
+void WriteVTP(cpp11::list inputTracts, const std::string &outputFile);
+[[cpp11::register]]
+void WriteFDS(cpp11::list inputTracts, const std::string &outputFile);
 
 #endif /* _FIBERWRITERS_H */