+++ tests/tinytest.R
@@ -1,0 +1,2 @@
+library(tinytest)
+test_package("riot")